test(deepdiff): Testcase that results in different output 6% of the time
These are two different datasets, one with an inlined body. The result of
diffing them usually gives expectA, but 6% of the time will give expectB.
The difference is that expectB returns everything being added in one insert,
whereas expectA correctly notes that "/meta/title" is updated, not inserted.
It's as if the empty body in dataset "left" sometimes causes deepdiff to
believe that nothing at all existed in the dataset.

This behavior needs to be deterministic so we can rely on deepdiff output
in unit tests in qri.