feat(address): reperesent address elements as strings, ints, or null
The original proposal I was hoping to use would be based heavily on the
joinpointer spec, which combines path elements into /slash/separated/strings.

In practice this is problematic for diff representation, which has to be able to
describe insertions into an array from, say a object. Using only strings for
addressing, this is ambiguous:

"14"

Is this address a string key in a map: "14", or an implict array index: 14?

By adding types to address elements, this ambiguity is removed, and we cut down
on some of the performance penalty of Constantly converting to and from strings.

While we're here, we might as well solve the other problem jsonpointer deals
with: referring to the root of a structured data tree. In jsonpointer, that's
referred to with the lone backslash: "/". For us, we'll use null, because we
have rich types. A path of null more accurately represents the root, which
has no path. This will also make it much easier to spot diffs that affect the
root, as the path elements will be null.

This means the only proper way to encode a deepdiff patch is with structured
data like JSON, which can distinguish between numeric, null and string types.
Although it's pretty easy to see that was already the case.