fix(diff): Prevent segfaults in bestCandidates
BestCandidates compares a list of candidate nodes, taken from the hashed nodes from d1, against a single node n2 from d2. During the comparison, nodes will be promoted to their parents if they don't immediately match. This can cause nil pointers to appear in the candidate list, and since the candidate list is aliased to the value of the t1Nodes map, over time multiple calls to BestCandidates may result in dereferencing a nil pointer, causing a segfault.

Instead, copy the candidate list, so it's safe to modify. Also, check if the candidate node is nil before trying to deref it. Note that this does not introduce the possibility of an infinite loop, since n2 is always promoted to its parent, which means it will eventually hit the root, and then break the loop.

I don't totally understand this function, so I can't write a unit test validating this behavior is improved. Would like to come back to deepdiff later and reevaluate its core algorithm.